Refactor to fix native selection and focus issues
See https://github.com/atom/atom/issues/16267

The Atom text editing components (text-editor, selection, marker) call
their event handlers synchronously. Within a handler of a cursor event
the text buffer, the cursor, and the selection cannot be reliably
updated (the editor loses focus somewhere during the updates). So the
updates were refactored to happen asynchronously. As a result the
cursor jump has become clearly visible, but this does not break the UX
of this package.